backend/requirements.txt
# ============================================================================
# Python Dependencies for Interactive AI Vision Assistant Backend
# ============================================================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI/ML Services
google-generativeai==0.3.2

# Speech-to-Text (Free Alternatives)
SpeechRecognition==3.10.0
pydub==0.25.1

# Text-to-Speech
gTTS==2.5.0
pyttsx3==2.90

# Session Management
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pydantic>=2.0.0,<3.0.0
Pillow>=10.0.0
